backend1/requirements.txt
Flask
Flask-Cors
opencv-python
numpy
shapely
anthropic
python-dotenv
networkx
